Ticket Revenue: Add public UI and refactor WPCS determination
Instead of using the currency to determine if a transaction went through
WPCS, this looks at the payment method meta data associated with the
transaction. This is probably more accurate than the currency method,
but it also introduces a new issue because some transactions don't have
a recorded payment method. It seems to only occur for transactions where
the total amount due was 0, either because the initial price of the
ticket was free, or a discount was applied that made the ticket free.
So, this also adds a new group to the compiled results, for transactions
without a payment method.